+++ provider/provider.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"context"
+	"embed"
 	"fmt"
 	"sync"
 	"sync/atomic"
@@ -38,21 +39,28 @@ type Provider struct {
 	Config func() Config
 	// Logger to call, this logger is passed to the serve.Serve Client, if not define Serve will create one instead.
 	Logger hclog.Logger
-	// Database connection
-	db schema.Database
+	// Database connection string
+	dbURL string
 	// meta is the provider's client created when configure is called
 	meta schema.ClientMeta
 	// Whether provider should all Delete on every table before fetching
 	disableDelete bool
 	// Add extra fields to all resources, these fields don't show up in documentation and are used for internal CQ testing.
 	extraFields map[string]interface{}
+	// Migrations embedded and passed by the provider to upgrade between versions
+	Migrations embed.FS
 }
 
 func (p *Provider) GetProviderSchema(_ context.Context, _ *cqproto.GetProviderSchemaRequest) (*cqproto.GetProviderSchemaResponse, error) {
+	m, err := readProviderMigrationFiles(p.Logger, p.Migrations)
+	if err != nil {
+		return nil, err
+	}
 	return &cqproto.GetProviderSchemaResponse{
 		Name:           p.Name,
 		Version:        p.Version,
 		ResourceTables: p.ResourceMap,
+		Migrations:     m,
 	}, nil
 }
 
@@ -64,24 +72,22 @@ func (p *Provider) GetProviderConfig(_ context.Context, _ *cqproto.GetProviderCo
 	data := fmt.Sprintf(`
 		provider "%s" {
 			%s
+			// list of resources to fetch
 			resources = %s
+			// enables partial fetching, allowing for any failures to not stop full resource pull
+			enable_partial_fetch = true
 		}`, p.Name, p.Config().Example(), helpers.FormatSlice(funk.Keys(p.ResourceMap).([]string)))
 
 	return &cqproto.GetProviderConfigResponse{Config: hclwrite.Format([]byte(data))}, nil
 }
 
-func (p *Provider) ConfigureProvider(ctx context.Context, request *cqproto.ConfigureProviderRequest) (*cqproto.ConfigureProviderResponse, error) {
+func (p *Provider) ConfigureProvider(_ context.Context, request *cqproto.ConfigureProviderRequest) (*cqproto.ConfigureProviderResponse, error) {
 	if p.meta != nil {
 		return &cqproto.ConfigureProviderResponse{Error: fmt.Sprintf("provider %s was already configured", p.Name)}, nil
 	}
-
-	conn, err := schema.NewPgDatabase(ctx, request.Connection.DSN)
-	if err != nil {
-		return nil, err
-	}
 	p.disableDelete = request.DisableDelete
-	p.db = conn
 	p.extraFields = request.ExtraFields
+	p.dbURL = request.Connection.DSN
 	providerConfig := p.Config()
 	if err := defaults.Set(providerConfig); err != nil {
 		return &cqproto.ConfigureProviderResponse{}, err
@@ -98,6 +104,14 @@ func (p *Provider) ConfigureProvider(ctx context.Context, request *cqproto.Confi
 	if err != nil {
 		return &cqproto.ConfigureProviderResponse{}, err
 	}
+
+	tables := make(map[string]string)
+	for r, t := range p.ResourceMap {
+		if err := getTableDuplicates(r, t, tables); err != nil {
+			return &cqproto.ConfigureProviderResponse{}, err
+		}
+	}
+
 	p.meta = client
 	return &cqproto.ConfigureProviderResponse{}, nil
 }
@@ -112,16 +126,29 @@ func (p *Provider) FetchResources(ctx context.Context, request *cqproto.FetchRes
 		return fmt.Errorf("provider has duplicate resources requested")
 	}
 
+	// if resources ["*"] is requested we will fetch all resources
+	resources, err := p.interpolateAllResources(request.Resources)
+	if err != nil {
+		return err
+	}
+
+	conn, err := schema.NewPgDatabase(ctx, p.Logger, p.dbURL)
+	if err != nil {
+		return fmt.Errorf("failed to connect to database. %w", err)
+	}
+
+	defer conn.Close()
+
 	g, gctx := errgroup.WithContext(ctx)
-	finishedResources := make(map[string]bool, len(request.Resources))
+	finishedResources := make(map[string]bool, len(resources))
 	l := sync.Mutex{}
 	var totalResourceCount uint64 = 0
-	for _, resource := range request.Resources {
+	for _, resource := range resources {
 		table, ok := p.ResourceMap[resource]
 		if !ok {
 			return fmt.Errorf("plugin %s does not provide resource %s", p.Name, resource)
 		}
-		execData := schema.NewExecutionData(p.db, p.Logger, table, p.disableDelete, p.extraFields)
+		execData := schema.NewExecutionData(conn, p.Logger, table, p.disableDelete, p.extraFields, request.PartialFetchingEnabled)
 		p.Logger.Debug("fetching table...", "provider", p.Name, "table", table.Name)
 		// Save resource aside
 		r := resource
@@ -136,15 +163,17 @@ func (p *Provider) FetchResources(ctx context.Context, request *cqproto.FetchRes
 			defer l.Unlock()
 			if err != nil {
 				return sender.Send(&cqproto.FetchResourcesResponse{
-					FinishedResources: finishedResources,
-					ResourceCount:     resourceCount,
-					Error:             err.Error(),
+					FinishedResources:           finishedResources,
+					ResourceCount:               resourceCount,
+					Error:                       err.Error(),
+					PartialFetchFailedResources: cqproto.PartialFetchToCQProto(execData.PartialFetchFailureResult),
 				})
 			}
 			err = sender.Send(&cqproto.FetchResourcesResponse{
-				FinishedResources: finishedResources,
-				ResourceCount:     resourceCount,
-				Error:             "",
+				FinishedResources:           finishedResources,
+				ResourceCount:               resourceCount,
+				Error:                       "",
+				PartialFetchFailedResources: cqproto.PartialFetchToCQProto(execData.PartialFetchFailureResult),
 			})
 			if err != nil {
 				return err
@@ -154,4 +183,34 @@ func (p *Provider) FetchResources(ctx context.Context, request *cqproto.FetchRes
 		})
 	}
 	return g.Wait()
+}
+
+func (p *Provider) interpolateAllResources(requestedResources []string) ([]string, error) {
+	if len(requestedResources) != 1 {
+		if funk.ContainsString(requestedResources, "*") {
+			return nil, fmt.Errorf("invalid \"*\" resource, with explicit resources")
+		}
+		return requestedResources, nil
+	}
+	if requestedResources[0] != "*" {
+		return requestedResources, nil
+	}
+	allResources := make([]string, 0, len(p.ResourceMap))
+	for k := range p.ResourceMap {
+		allResources = append(allResources, k)
+	}
+	return allResources, nil
+}
+
+func getTableDuplicates(resource string, table *schema.Table, tableNames map[string]string) error {
+	for _, r := range table.Relations {
+		if err := getTableDuplicates(resource, r, tableNames); err != nil {
+			return err
+		}
+	}
+	if existing, ok := tableNames[table.Name]; ok {
+		return fmt.Errorf("table name %s used more than once, duplicates are in %s and %s", table.Name, existing, resource)
+	}
+	tableNames[table.Name] = resource
+	return nil
 }