+++ provider/schema/database.go
@@ -7,10 +7,14 @@ import (
 	"reflect"
 	"strconv"
 
+	"github.com/hashicorp/go-hclog"
+	"github.com/jackc/pgconn"
+
 	"github.com/modern-go/reflect2"
 
 	"github.com/doug-martin/goqu/v9"
 
+	"github.com/jackc/pgerrcode"
 	"github.com/jackc/pgx/v4"
 
 	sq "github.com/Masterminds/squirrel"
@@ -30,13 +34,15 @@ type Database interface {
 	Delete(ctx context.Context, t *Table, args []interface{}) error
 	Query(ctx context.Context, query string, args ...interface{}) (pgx.Rows, error)
 	CopyFrom(ctx context.Context, resources Resources, shouldCascade bool, CascadeDeleteFilters map[string]interface{}) error
+	Close()
 }
 
 type PgDatabase struct {
 	pool *pgxpool.Pool
+	log  hclog.Logger
 }
 
-func NewPgDatabase(ctx context.Context, dsn string) (*PgDatabase, error) {
+func NewPgDatabase(ctx context.Context, logger hclog.Logger, dsn string) (*PgDatabase, error) {
 	cfg, err := pgxpool.ParseConfig(dsn)
 	if err != nil {
 		return nil, err
@@ -45,7 +51,7 @@ func NewPgDatabase(ctx context.Context, dsn string) (*PgDatabase, error) {
 	if err != nil {
 		return nil, err
 	}
-	return &PgDatabase{pool: pool}, nil
+	return &PgDatabase{pool: pool, log: logger}, nil
 }
 
 // Insert inserts all resources to given table, table and resources are assumed from same table.
@@ -70,6 +76,10 @@ func (p PgDatabase) Insert(ctx context.Context, t *Table, resources Resources) e
 
 	s, args, err := sqlStmt.ToSql()
 	if err != nil {
+		// This should rarely occur, but if it does we want to print the SQL to debug it further
+		if pgErr, ok := err.(*pgconn.PgError); ok && pgerrcode.IsSyntaxErrororAccessRuleViolation(pgErr.Code) {
+			p.log.Error("insert syntax error", "sql", s)
+		}
 		return err
 	}
 	_, err = p.pool.Exec(ctx, s, args...)
@@ -153,6 +163,10 @@ func (p PgDatabase) Delete(ctx context.Context, t *Table, args []interface{}) er
 
 	_, err = p.pool.Exec(ctx, sql, args...)
 	return err
+}
+
+func (p PgDatabase) Close() {
+	p.pool.Close()
 }
 
 func GetPgTypeFromType(v ValueType) string {